move tanglit command to tanglit module
